apply fast sets to NodeSetMod, NodeSetDivGreedy NodeSetModOpt2: add partitionLouvain(), recursiveLouvain(): slightly better modularity

+++ uncertain-graph-java/src/algs4/EdgeInt.java
@@ -57,7 +57,7 @@ public class EdgeInt implements Comparable<EdgeInt> {
      * Returns the weight of the edge.
      * @return the weight of the edge
      */
-    public double weight() {
+    public int weight() {
         return weight;
     }
